requirements.txt
aiobotocore==2.4.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
altair==4.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
botocore==1.27.59
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
lightgbm==3.3.3
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
meteostat==1.6.5
multidict==6.0.2
nest-asyncio==1.5.6
numba==0.56.2
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
s3fs==2022.8.2
scikit-learn==1.0.2
scipy==1.7.3
semver==2.13.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
streamlit==1.13.0
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing-extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
zipp==3.9.0
